requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
datasets==3.5.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
multidict==6.4.3
multiprocess==0.70.16
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.2.1
polars==1.27.1
propcache==0.3.1
pyarrow==19.0.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
